requirements.txt
functions-framework
flask
openai
python-dotenv